# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "tags" : [ {
    "name" : "Users API",
    "description" : "A simple API for managing users"
  }, {
    "name" : "Activities API",
    "description" : "A simple API for managing activities"
  }, {
    "name" : "Timelogs API",
    "description" : "A simple API for managing timelogs"
  }, {
    "name" : "Projects API",
    "description" : "A simple API for managing projects"
  } ],
  "paths" : {
    "/api/v1/project/{id}" : {
      "get" : {
        "tags" : [ "Projects API" ],
        "summary" : "Get a single project",
        "operationId" : "getProjectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Projects API" ],
        "summary" : "Update an existing project",
        "operationId" : "updateProject",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity/{id}" : {
      "get" : {
        "tags" : [ "Activities API" ],
        "summary" : "Get a single activity",
        "operationId" : "getActivityById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Activities API" ],
        "summary" : "Update an existing activity",
        "operationId" : "updateActivity",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ActivityDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/timelog/{activityId}" : {
      "post" : {
        "tags" : [ "Timelogs API" ],
        "summary" : "Save a timelog",
        "operationId" : "saveTimeLog",
        "parameters" : [ {
          "name" : "activityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TimeLogDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeLogDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity/{id}/timelog" : {
      "post" : {
        "tags" : [ "Activities API" ],
        "summary" : "Add TimeLog to activity",
        "operationId" : "addTimeLogToActivity",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TimeLogDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user" : {
      "get" : {
        "tags" : [ "Users API" ],
        "summary" : "Get all available users",
        "operationId" : "getAllUsers",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/{id}" : {
      "get" : {
        "tags" : [ "Users API" ],
        "summary" : "Get a single user",
        "operationId" : "getUserById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/timelog" : {
      "get" : {
        "tags" : [ "Timelogs API" ],
        "summary" : "Get all available timelogs",
        "operationId" : "getAllTimeLogs",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TimeLogDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/project" : {
      "get" : {
        "tags" : [ "Projects API" ],
        "summary" : "Get all available projects",
        "operationId" : "getAllProjects",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/project/user/{id}" : {
      "get" : {
        "tags" : [ "Projects API" ],
        "summary" : "Get projects with a specified user",
        "operationId" : "getProjectsByUserId",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectsByUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity" : {
      "get" : {
        "tags" : [ "Activities API" ],
        "summary" : "Get all available activities",
        "operationId" : "getAllActivities",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ActivityDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/timelog/{id}" : {
      "delete" : {
        "tags" : [ "Timelogs API" ],
        "summary" : "Delete a timelog",
        "operationId" : "deleteTimeLog",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ActivityDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "timeLogs" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimeLogDto"
            }
          }
        }
      },
      "ProjectDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "activities" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ActivityDto"
            }
          }
        }
      },
      "TimeLogDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "duration" : {
            "type" : "string"
          },
          "isReported" : {
            "type" : "boolean"
          },
          "isSubmitted" : {
            "type" : "boolean"
          }
        }
      },
      "UserDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "firstName" : {
            "type" : "string",
            "description" : "The users first name",
            "example" : "Bob"
          },
          "lastName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        }
      },
      "ProjectsByUserDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "projects" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProjectDto"
            }
          }
        }
      }
    }
  }
}